Remove DraftTransformer and change packages of CategorySync.
